+++ CHANGELOG.md
@@ -4,3 +4,5 @@
   - List support while editing
   - List support fixes in `CreateEntryBottomSheet`
 - Motion fixes in `DetailEntryBottomSheet`
+- Search
+  - Search Data Layer